Updated the API to pass the statuses along to the client with a viewmodel.
Added form validation to the customer edit form.

Enabled editing of the notes and adding new notes.

altered the css to enable compact rows for segmented collections.

modified the css to properly display a full height flex page with a sticky footer.

Renamed the rxjs import file to Operators.

removed the animations import code until it is more stable.

removed some testing code as well as some extra config files not used with this project.